Documentacion de las bibliotecas

+++ Tp1/src/utn_calculos.h
@@ -12,6 +12,6 @@ int utn_getSuma (int operador1, int operador2, int*resultado,char*mensaje, char*
 int utn_getResta (int operador1, int operador2, int*resultado,char*mensaje, char*mensajeError);
 float utn_getDividir (int operador1, int operador2,float* resultado, char*mensaje, char*mensajeError);
 int utn_getMultiplicar (int operador1, int operador2, int*resultado, char*mensaje, char*mensajeError);
-int utn_getFactorial (int operador, int * resultado , char*mensaje, char*mensajeError);
+int utn_getFactorial (int operador, double* resultado , char*mensaje, char*mensajeError);
 
 #endif /* UTN_CALCULOS_H_ */
